Misc bugfixes and enhancements and tidy-ups.
Bugfix: if BB was installed in /, the path to /directory/server would be //directory/server; the browser tried to connect to http://directory/server. This appeared in a couple of places. Now //'s are replaced with a single slash.

getchannels() is now called _getchannels to block outside calls.

Added _getregistrations(), so this module no longer depends on Sofia.

Simplified a switch to an if/elseif, removed some unused cases.

removed test page test().