fairly major reorganisation for directory.
views/directory/index.php was doubling as conferences, this is now split off.

The ajax fetch used to read from the database, now it only reads from freeswitch.

This allows a hypothetical piece of code to be written which calls Directory_Controller::json_listing, which bypasses 99% of the framework, if the framework is too heavy on a particular (hypothetical) low-powered box.