Documentation on how to get a 10-fold speed boost. Ooh-eck!
Using Kohana adds a significant load. For the directory listing, this may be too much for an embedded box like mine, if it's being polled regularly by several people.

I have created a patch to circumvent Kohana, for just this page. It is an ugly kludge, circumvents BB security, etc., so I'm putting it in as documentation how-to. Perhaps in the future there will be a good way of doing this.

If you really think you need this, and are willing to accept the consequences, then read the README file.