Fix Multiple regex's in simpleroute
See http://tickets.2600hz.org/browse/BLUEBOX-228

Combining regexes (regii?) is *very* complicated, if not impossible. This gets around the problem by leaving them seperate, and having multiple rules to compensate.

Also, if you put in a non-regex pattern with brackets, it puts brackets around the entire thing. The number that is dialed is always $1, so it doesn't make any sense to have a pattern of '1234' - it will forward the call with a number of ''.