Lots of little developements and tweaks
in function config, provisioner_lib->options is set up at the start, so it can be changed throughout the function

clean up the setting of $provisioner_lib->DateTimeZone

For snoms only, set the tone_scheme - this should be generalized later.

New function _autoquestion - ask a question from a make & model of phone. It is up to the calling code to determine
which questions should be asked.

In function settings() - ask about the snom default tone scheme.

in _getprovisioningdata() -  the "data" tag is a list (multi choice)