More work on endpointmanager
Add dateformat (defaults to little-endian) and tonescheme (defaults to USA); add cisco image_name, to specify the firmware. Defaults to blank, so it doesn't specify any firmware, which the phone seems to be fine with.

Remove reference to $provisioner_lib->options, which never seems to get used.

The code to read in all the json data from the provisioner library seperated the families so each model stood independently - in hindsight, this was a mistake. Put the families back in. Autoquestion now refers to families, not models.
Also read in json for global_template_data.json.

Added global autoquestions - this is a much quicker way of adding questions, by pulling the info from the provisioner.

moved some questions from snom to global section, since they are repeated for all phones.

in function dump(), set the content_header AFTER getting provisioningdata, so if getting provisioning data fails, the html is shown as html.