Misc changes to endpointmanager:
Put _identify_configfile_regex back in - there are some files which need to be identified by regex, rather than mac, since they do not have a mac address in them. This is a work in progress, and is disabled at the moment. modified result of identify_configfile (i.e. non regex version) to match the regex version, so hopefully they will be interchangeable, somewhat.

Set a default timeformat (24 hour)

fix name of snom's admin_pass to admin_mode_password.

Change default of cisco's preferredcodec to 'none' - this means it negotiates with the server, which should be correct in most cases.

Ask for the firmware for various models. List of firmware is taken by a glob, matched to the firmwares directory.

Added lookup from brand/model to family.

Add an option to config to allow file specified as config?file=uriescapedconfigfilename.

Set the content-type for all config files as text/plain. This means that xml files are visible.

Replace "make" with "brand", when referring to phones. Instead of a mixture.

If an autoquestion cannot be found, show an error message.

The section names for the different brands of phone are now in alphabetical order.

_jsonread now does a lot of checking, where checking is available, to show as helpfull a message as possible.

rejigger of _getprovisioningdata