Fix error message in xmlcdr's xml report
Before this fix, if you exported the call records as XML (the "else" clause"), ob_end_clean would be called, disabling the buffer. Then, immediately after the end of the "else" clause, ob_end_flush would be called. Since the buffer was disabled, ob_end_flush would generate an error message, which would be tacked on to the end of the xml report.

This change fixes this by replacing ob_end_clean with ob_clean, which also deletes the current contents of the buffer, but does not disable buffering.